+++ frontend_react/react_src/src/constants/dev.ts
@@ -1,1 +1,0 @@
-export const VITE_DEFAULT_PORT = 5173;